Fix Type mismatch after KT-56583

+++ ktor-client/ktor-client-winhttp/windows/src/io/ktor/client/engine/winhttp/internal/WinHttpExceptions.kt
@@ -96,7 +96,7 @@ private fun formatMessage(errorCode: UInt, moduleHandle: HMODULE? = null): Strin
         errorCode,
         languageId,
         bufferPtr.reinterpret(),
-        0,
+        0.toUInt(),
         null
     )
